extracted enums/classes to own files
