+++ src/gverse/debug-logger.ts
@@ -1,0 +1,8 @@
+import debug from "debug"
+
+/* To see Gverse debug logs, run the scrpit or test, set environment variable:
+ * `DEBUG=Gverse <script ...>`
+ */
+const log = debug("Gverse")
+
+export default log
